@@ -1,24 +1,28 @@
-source 'https://rubygems.org'
+source 'http://rubygems.org'
 
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '4.0.2'
+#gem 'rails', '4.0.2'
+gem 'rails', '~> 5.2'
+
+gem 'railties', '5.2.0'
 
 # Use sqlite3 as the database for Active Record
 #gem 'sqlite3'
-gem 'pg'
+
+gem 'pg', '~> 0.18'
 
 gem 'cocoon'
 
 gem "font-awesome-rails"
 
 # Use SCSS for stylesheets
-gem 'sass-rails', '~> 4.0.0'
+gem 'sass-rails'
 
 # Use Uglifier as compressor for JavaScript assets
-gem 'uglifier', '>= 1.3.0'
+gem 'uglifier'
 
 # Use CoffeeScript for .js.coffee assets and views
-gem 'coffee-rails', '~> 4.0.0'
+gem 'coffee-rails'
 
 # See https://github.com/sstephenson/execjs#readme for more supported runtimes
 # gem 'therubyracer', platforms: :ruby
@@ -30,7 +34,7 @@ gem 'jquery-rails'
 gem 'turbolinks'
 
 # Build JSON APIs with ease. Read more: https://github.com/rails/jbuilder
-gem 'jbuilder', '~> 1.2'
+gem 'jbuilder'
 
 group :doc do
   # bundle exec rake doc:rails generates the API under doc/api.
@@ -44,10 +48,14 @@ end
  gem 'unicorn'
 
 # Use Capistrano for deployment
-gem 'capistrano', '2.13.5'
+gem 'capistrano'
 
 # gem 'momentjs-rails', '~> 2.5.0'
 # gem 'bootstrap3-datetimepicker-rails', '~> 3.0.0'
 
 # Use debugger
 # gem 'debugger', group: [:development, :test]
+
+gem 'geocoder'
+
+gem 'gmaps4rails'
