@@ -1,4 +1,4 @@
-class CreateSchools < ActiveRecord::Migration
+class CreateSchools < ActiveRecord::Migration[5.2]
   def change
     create_table :schools do |t|
       t.string :name
