@@ -1,4 +1,4 @@
-class CreateResidences < ActiveRecord::Migration
+class CreateResidences < ActiveRecord::Migration[5.2]
   def change
     create_table :residences do |t|
       t.string :name
