@@ -1,4 +1,4 @@
-class CreateResidenceFeatures < ActiveRecord::Migration
+class CreateResidenceFeatures < ActiveRecord::Migration[5.2]
   def change
     create_table :residence_features do |t|
       t.integer :residence_id
