@@ -1,4 +1,4 @@
-class CreateLnkSchoolCourses < ActiveRecord::Migration
+class CreateLnkSchoolCourses < ActiveRecord::Migration[5.2]
   def change
     create_table :lnk_school_courses do |t|
       t.integer :school_id
