@@ -1,4 +1,4 @@
-class CreateCourses < ActiveRecord::Migration
+class CreateCourses < ActiveRecord::Migration[5.2]
   def change
     create_table :courses do |t|
       t.string :name
