@@ -1,4 +1,4 @@
-class CreateLangLevels < ActiveRecord::Migration
+class CreateLangLevels < ActiveRecord::Migration[5.2]
   def change
     create_table :lang_levels do |t|
       t.string :name
