@@ -1,4 +1,4 @@
-class CreateCurrencies < ActiveRecord::Migration
+class CreateCurrencies < ActiveRecord::Migration[5.2]
   def change
     create_table :currencies do |t|
       t.string :code
