@@ -1,4 +1,4 @@
-class CreateExchangeRates < ActiveRecord::Migration
+class CreateExchangeRates < ActiveRecord::Migration[5.2]
   def change
     create_table :exchange_rates do |t|
       t.integer :currency_1_id
