@@ -10,6 +10,24 @@ require 'csv'
 
 Company.delete_all
 CSV.foreach("companies.csv", :headers => true) do |row|
-  #puts row['CD компании']+row['Краткое название']
+  puts row['CD компании']+row['Краткое название']
   Company.create!(:name => row['Полное название'], :description => row['Описание '])
+end
+
+School.delete_all
+CSV.foreach("seeds/dump_schools.csv", :headers => true) do |row|
+  puts row.to_hash
+  School.create!(row.to_hash)
+end
+
+Course.delete_all
+CSV.foreach("seeds/dump_courses.csv", :headers => true) do |row|
+  puts row.to_hash
+  Course.create!(row.to_hash)
+end
+
+Language.delete_all
+CSV.foreach("seeds/dump_languages.csv", :headers => true) do |row|
+  puts row.to_hash
+  Language.create!(row.to_hash)
 end
